lambda_functions/aviationedge/requirements.txt
boto3==1.35.55
requests==2.32.3
python-dotenv==1.0.1
pyyaml==6.0.2
psycopg2-binary==2.9.10